core/serializers/business_serializers.py
# core/serializers/business_serializers.py

from rest_framework import serializers
from ..models import Business, Table

class BusinessSerializer(serializers.ModelSerializer):
    """
    İşletme modelinin detaylarını serileştirmek için kullanılır.
    """
    class Meta:
        model = Business
        # === GÜNCELLEME BURADA: 'timezone' alanı fields listesine eklendi ===
        fields = ['id', 'owner', 'name', 'address', 'phone', 'is_setup_complete', 'currency_code', 'timezone']


class TableSerializer(serializers.ModelSerializer):
    """
    Masa modelini serileştirmek için kullanılır.
    """
    class Meta:
        model = Table
        fields = ['id', 'table_number', 'uuid', 'business']