requirements.txt
anyio==4.9.0
asgiref==3.8.1
bidict==0.23.1
channels==4.2.2
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2024.1                   # Eklenen kütüphane
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.34.0
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0

# Google API istemcisi ve kimlik doğrulama
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# PostgreSQL için
psycopg[binary]
dj-database-url==2.1.0

# Heroku için önerilenler
gunicorn==22.0.0
whitenoise==6.6.0
daphne==4.1.0